icE1usb fw: Use 'int' as the default type for port index
No benefit to use uint8_t and for consistency with later patches,
use 'int' here too.

+++ firmware/ice40-riscv/icE1usb/e1.h
@@ -57,5 +57,5 @@ enum e1_platform_led_state {
 	E1P_LED_ST_BLINK_FAST	= 3
 };
 
-extern void e1_platform_led_set(uint8_t port, enum e1_platform_led led,
+extern void e1_platform_led_set(int port, enum e1_platform_led led,
 				enum e1_platform_led_state state);
